requirements.txt
mysql-connector-python
pandas
openpyxl
pillow
reportlab
